requirements.txt
langchain
python-dotenv
langchain-openai
langchain-community
bs4
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain_huggingface
langchain_groq
langgraph
langgraph-cli[inmem]
fastapi
uvicorn
langchain-groq
langchain-core
langgraph
python-dotenv